0.8

+++ zsh_athame_setup.sh
@@ -25,12 +25,15 @@ athame=1
 dirty=0
 rc=1
 submodule=1
+sudo="sudo "
 vimbin=""
 destdir=""
+docdir=""
+htmldir=""
 prefix="/usr"
 for arg in "$@"
 do
-  case $arg in
+  case "$arg" in
     "--redownload" ) redownload=1;;
     "--nobuild" ) build=0;;
     "--noathame" ) athame=0;;
@@ -38,6 +41,7 @@ do
     "--dirty" ) dirty=1;;
     "--norc" ) rc=0;;
     "--nosubmodule" ) submodule=0;;
+    "--nosudo" ) sudo="";;
     --vimbin=*) vimbin="${arg#*=}";;
     --destdir=*) destdir="${arg#*=}";;
     --prefix=*) prefix="${arg#*=}";;
@@ -65,9 +69,9 @@ do
 done
 
 prefix_flag="--prefix=$prefix"
-libdir_flag=${libdir:+"--libdir=$libdir"}
-docdir_flag="--docdir=${docdir-$prefix/share/doc/zsh}"
-htmldir_flag="--htmldir=${htmldir-$docdir/html}"
+libdir_flag="${libdir:+--libdir=$libdir}"
+docdir_flag="--docdir=${docdir:-$prefix/share/doc/zsh}"
+htmldir_flag="--htmldir=${htmldir:-$docdir/html}"
 
 #Get vim binary
 if [ -z $vimbin ]; then
@@ -99,11 +103,11 @@ if [ $redownload = 1 ]; then
 fi
 if [ ! -f zsh-5.1.1.tar.gz ]; then
   curl -O http://www.zsh.org/pub/old/zsh-5.1.1.tar.gz
-  if [ "$(md5sum zsh-5.1.1.tar.gz)" != "8ba28a9ef82e40c3a271602f18343b2f  zsh-5.1.1.tar.gz" ]; then
-    rm zsh-5.1.1.tar.gz
-    echo "FAILED: Incorrect md5 hash" >&2
-    exit 1
-  fi
+fi
+if [ "$(md5sum zsh-5.1.1.tar.gz 2>/dev/null)" != "8ba28a9ef82e40c3a271602f18343b2f  zsh-5.1.1.tar.gz" ] && [ "$(md5 zsh-5.1.1.tar.gz 2>/dev/null)" != "MD5 (zsh-5.1.1.tar.gz) = 8ba28a9ef82e40c3a271602f18343b2f" ]; then
+  rm zsh-5.1.1.tar.gz
+  echo "FAILED: Incorrect md5 hash" >&2
+  exit 1
 fi
 
 if [ ! -d zsh-5.1.1_tmp ]; then
@@ -133,9 +137,10 @@ fi
 #Build and install zsh
 if [ $build = 1 ]; then
   if [ ! -f Makefile ]; then
-    ./configure $prefix_flag \
-        $docdir_flag \
-        $htmldir_flag \
+    ./configure "$prefix_flag" \
+        "$libdir_flag" \
+        "$docdir_flag" \
+        "$htmldir_flag" \
         --enable-etcdir=/etc/zsh \
         --enable-zshenv=/etc/zsh/zshenv \
         --enable-zlogin=/etc/zsh/zlogin \
@@ -143,11 +148,10 @@ if [ $build = 1 ]; then
         --enable-zprofile=/etc/zsh/zprofile \
         --enable-zshrc=/etc/zsh/zshrc \
         --enable-maildir-support \
-        --with-term-lib='ncursesw' \
         --enable-multibyte \
         --enable-function-subdirs \
-        --enable-fndir=/usr/share/zsh/functions \
-        --enable-scriptdir=/usr/share/zsh/scripts \
+        --enable-fndir="$prefix/share/zsh/functions" \
+        --enable-scriptdir="$prefix/share/zsh/scripts" \
         --with-tcsetpgrp \
         --enable-pcre \
         --enable-cap \
@@ -155,46 +159,54 @@ if [ $build = 1 ]; then
         || exit 1
   fi
   if [ $runtest = 1 ]; then
-    rm -rf $(pwd)/../test/build
-    mkdir -p $(pwd)/../test/build
+    rm -rf "$(pwd)/../test/build"
+    mkdir -p "$(pwd)/../test/build"
 
     # make sure the files affected by ATHAME_TESTDIR are updated to use test settings
     rm -f Src/zshpaths.h && touch Src/Zle/athame.c
 
-    mkdir -p $(pwd)/../test/build/usr/lib
-    make ATHAME_VIM_BIN=$vimbin ATHAME_TESTDIR=$(pwd)/../test/build || exit 1
-    make install DESTDIR=$(pwd)/../test/build || exit 1
+    make ATHAME_VIM_BIN="$vimbin" ATHAME_TESTDIR="$(pwd)/../test/build" || exit 1
+    make install DESTDIR="$(pwd)/../test/build" || exit 1
 
     # make sure the files affected by ATHAME_TESTDIR are updated to not use test settings
     rm -f Src/zshpaths.h && touch Src/Zle/athame.c
 
     cd ../test
-    ./runtests.sh "script -c ../build/usr/bin/zsh" || exit 1
+    zsh_bin="$(find $(pwd)/build -name zsh -type f | head -n 1)"
+    if [ "$(uname)" == "Darwin" ]; then
+      ./runtests.sh "script /dev/null $zsh_bin" || exit 1
+    else
+      ./runtests.sh "script -c $zsh_bin" || exit 1
+    fi
     cd -
-    make ATHAME_VIM_BIN=$vimbin || exit 1
+    make ATHAME_VIM_BIN="$vimbin" || exit 1
     echo "Installing Zsh with Athame..."
     if [ -n "$destdir" ]; then
-      mkdir -p $destdir
+      mkdir -p "$destdir"
     fi
     if [ -w "$destdir" ]; then
-      make install DESTDIR=$destdir || exit 1
+      make install DESTDIR="$destdir" || exit 1
     else
-      sudo make install DESTDIR=$destdir || exit 1
+      ${sudo}make install DESTDIR="$destdir" || exit 1
     fi
   else
     make ATHAME_VIM_BIN=$vimbin || exit 1
     echo "Installing Zsh with Athame..."
     if [ -n "$destdir" ]; then
-      mkdir -p $destdir
+      mkdir -p "$destdir"
     fi
     if [ -w "$destdir" ]; then
-      make install DESTDIR=$destdir || exit 1
+      make install DESTDIR="$destdir" || exit 1
     else
-      sudo make install DESTDIR=$destdir || exit 1
+      ${sudo}make install DESTDIR="$destdir" || exit 1
     fi
   fi
 fi
 
 if [ $rc = 1 ]; then
-  sudo cp ../athamerc /etc/athamerc
+  if [ -z "$sudo" ]; then
+    printf "\e[0;31mThe athamerc was not copied. You should copy athamerc to /etc/athamerc or ~/.athamerc.\e[0;0m\n"
+  else
+    sudo cp ../athamerc /etc/athamerc
+  fi
 fi
